Pre-cache releases data
